p2p: Adds helper function to generate the base reconciliation sketch
Sketch will be computed on the fly every tricking interval. Data to be added to
the sketch is added to a temporary collection so the node cannot be probed in between
intervals. When the interval is reached, the data is added to the sketch and the temp
collections is cleared.

Sketched are also initiated at max size, and will be shrinked to size when exchanged.