Attach consider_fanout flag to mempool and orphanage transactions
When scheduling the relay of a transaction (RelayTransaction) we should consider
whether it is worth fanning it out, or only reconciling it. This depends, partially,
on how the transaction was received.

Since the transaction reception and the relay scheduling do not happen on the same call,
we need to store this flag on transaction may be relayed (those stored in the mempool and
the orphanage).

FIXME: Currently, we are not properly passing this flag to transaction on the orphanage nor
to packages, the flag is hardcoded in both cases. Make sure this is how we want to persist the
flag address both cases.