txrequest: adds m_first_inv to TxRequestTracker and reworks ReceivedInv
Keeps track of the first received inv for a given transaction in txrequest.
All matching txids will share the same first inv time.

Also reworks TxRequestTracker::ReceivedInv to receive a recvtime and a delay
instead of a reqtime(=recvtime + delay). This way, m_first_inv can be set
to recvtime (when applicable), and reqtime can be computed internally instead of
externally.